camera_control/requirements.txt
numpy
Pillow
supervision
ultralytics
